resolved comments
